feat: Add Statuses management (#10)
* feat: add Statuses

* feat: add ordering by id to all repositories

* fix: add admin permission checks in genres_api

+++ backend/api/v1/genres_api.py
@@ -11,6 +11,7 @@ from utils.db_utils import get_db_session
 from excepitons.genre_exceptions import DeleteGenreException
 from repositories.genre_repository import GenreRepository
 from services import authenticate
+from permissions import IsAdmin
 
 from schemas.genre_schema import (
     GenrePreview,
@@ -38,6 +39,7 @@ async def retrieve_genre(genre_id: int,
 async def create_genre(request: Request,
                        create_data: GenreCreate,
                        session: Session = Depends(get_db_session)):
+    IsAdmin.check_permissions(request)
     return GenreRepository.create_genre(session, create_data)
 
 
@@ -47,6 +49,7 @@ async def partial_update_genre(request: Request,
                                genre_id: int,
                                partial_update_data: GenreUpdate = Body(...),
                                session: Session = Depends(get_db_session)):
+    IsAdmin.check_permissions(request)
     return GenreRepository.partial_update_genre(
         session, genre_id, partial_update_data
     )
@@ -57,6 +60,8 @@ async def partial_update_genre(request: Request,
 async def delete_genre(request: Request,
                        genre_id: int,
                        session: Session = Depends(get_db_session)):
+    IsAdmin.check_permissions(request)
+
     is_deleted = GenreRepository.delete_genre(session, genre_id)
 
     if is_deleted:
